Refactor and build basic echo ws endpoint
