+++ backend/yagram-core/src/app_data/mod.rs
@@ -1,0 +1,3 @@
+mod app_state;
+
+pub use app_state::*;
