+++ backend/yagram-core/src/app_settings/mod.rs
@@ -1,0 +1,3 @@
+mod settings_initializer;
+
+pub use settings_initializer::*;
