+++ backend/yagram-core/src/domains/messages/mod.rs
@@ -1,0 +1,1 @@
+pub mod handlers;
