+++ backend/yagram-core/src/domains/mod.rs
@@ -1,2 +1,3 @@
 pub mod auth;
+pub mod messages;
 pub mod users;
