+++ backend/yagram-core/src/domains/users/mutations.rs
@@ -8,7 +8,7 @@ use sea_orm::InsertResult;
 pub async fn create_user_from_auth0(
     db: &DbConn,
     data: Auth0UserInfo,
-) -> Result<InsertResult<user::ActiveModel>, DbErr> {
+) -> Result<user::Model, DbErr> {
     let mut new_user: user::ActiveModel = Default::default();
     new_user.set(user::Column::FirstName, data.given_name.into());
     new_user.set(user::Column::LastName, data.family_name.into());
@@ -17,5 +17,5 @@ pub async fn create_user_from_auth0(
     new_user.set(user::Column::Email, data.email.into());
     new_user.set(user::Column::CreatedAt, Utc::now().naive_utc().into());
 
-    user::Entity::insert(new_user).exec(db).await
+    new_user.insert(db).await
 }
